[refactor-first-hit] pay no attention to the man behind the camera
The hypothetical ray gets extended in both directions: both behind and
in front of the "camera".  So I have to ignore everything "behind" it.
By convention, this is everything with a positive z-coordinate (which is
less elegant than I'd like).

As you may have noticed, I started writing better-formatted commit
messages, now that I'm using CIDER and Magit.