[bugfix-intersect-issues] comment no longer applies
The comment is no longer applicable.  It turns-out the fix was even
simpler than was implied in the comment.  Since we only care about
positive values of the parameter, corresponding to the intersection
points of the ray with the shape in front of the camera (solutions
behind the camera make no sense), we can discard negative roots.