[bugfix/first-hit] another test
Added a second test for first-hit.  It's still pretty basic, so I'm not
sure how much it proves.  The real test will come when I start trying to
render a slightly more complex scene.  I should probably craft a test
that has more than one object in the path of the test ray.