[feature/finishing-touches] world needs to be flat
I inadvertently created a nested list for the scene description, which
is why the code barfed.  It expects the list to be flat (1-D).

The scene from the book renders properly now.  Though the x and y axes
are transposed.  The fix ought to be a trivial one, of convention.